refining Type (working state)

+++ src/Class.cpp
@@ -27,7 +27,7 @@ Class::Class(
     int64_t id,
     std::string name,
     double priority,
-    int default_maxspeed
+    double default_maxspeed
     )
 :
   id(id),
